Lecture 324. Serving images statically

+++ controllers/admin.js
@@ -1,5 +1,6 @@
-const { validationResult } = require('express-validator/check')
+const mongoose = require('mongoose');
 
+const { validationResult } = require('express-validator/check');
 const errorHandlerObjectWrapper = require('../util/errorHandlerObjectWrapper');
 
 const Product = require('../models/product');
@@ -25,24 +26,33 @@ exports.postAddProduct = (req, res, next) => {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
       editing: false,
-      product: { title, price, description },
       hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
       errorMessage: 'Attached file is not an image.',
       validationErrors: [],
     });
   }
-  
-  const { errors } = validationResult(req);
+  const errors = validationResult(req);
 
-  if (errors.length > 0) {
+  if (!errors.isEmpty()) {
+    console.log(errors.array());
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
       editing: false,
-      product: { title, price, description },
       hasError: true,
-      errorMessage: errors[0].msg,
-      validationErrors: errors,
+      product: {
+        title: title,
+        imageUrl: imageUrl,
+        price: price,
+        description: description,
+      },
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
     });
   }
 
@@ -67,13 +77,20 @@ exports.postAddProduct = (req, res, next) => {
       //   pageTitle: 'Add Product',
       //   path: '/admin/add-product',
       //   editing: false,
-      //   product: { title, imageUrl, price, description },
       //   hasError: true,
-      //   errorMessage: 'Database operation failed.',
-      //   validationErrors: [],
+      //   product: {
+      //     title: title,
+      //     imageUrl: imageUrl,
+      //     price: price,
+      //     description: description
+      //   },
+      //   errorMessage: 'Database operation failed, please try again.',
+      //   validationErrors: []
       // });
       // res.redirect('/500');
-      return errorHandlerObjectWrapper(500, err);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
 
@@ -107,22 +124,23 @@ exports.postEditProduct = (req, res, next) => {
   const updatedPrice = req.body.price;
   const image = req.file;
   const updatedDesc = req.body.description;
-  const { errors } = validationResult(req);
 
-  if (errors.length > 0) {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
       path: '/admin/edit-product',
       editing: true,
+      hasError: true,
       product: {
         title: updatedTitle,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId,
       },
-      hasError: true,
-      errorMessage: errors[0].msg,
-      validationErrors: errors,
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
     });
   }
 
@@ -137,11 +155,10 @@ exports.postEditProduct = (req, res, next) => {
       if (image) {
         product.imageUrl = image.path;
       }
-      return product.save()
-        .then(result => {
-          console.log('UPDATED PRODUCT!');
-          res.redirect('/admin/products');
-        });
+      return product.save().then(result => {
+        console.log('UPDATED PRODUCT!');
+        res.redirect('/admin/products');
+      });
     })
     .catch(err => errorHandlerObjectWrapper(500, err));
 };
@@ -163,10 +180,7 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({
-    _id: prodId,
-    userId: req.user._id,
-  })
+  Product.deleteOne({ _id: prodId, userId: req.user._id })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
